requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pytorch>=1.12.0
python-multipart==0.0.6
openai-whisper==20230314